+++ src/main/java/instruments/Trumpet.java
@@ -1,12 +1,10 @@
 package instruments;
 
-import behaviours.IPlay;
-
 public class Trumpet extends Instrument {
 
     private int numOfValves;
 
-    public Trumpet(String type, String colour, int stockPrice, int sellPrice, int numOfValves){
+    public Trumpet(String type, String colour, double stockPrice, double sellPrice, int numOfValves){
         super("brass", colour, stockPrice, sellPrice);
         this.numOfValves = numOfValves;
     }
